Moved window keywords to own component
